+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## NEXT RELEASE
 
 - #509: Change/Fix to use a symmetric barrier to synchronize domains
+- #511: Introduce extended specs to allow wrapping command sequences
 
 ## 0.6